+++ js/data.js
@@ -1,0 +1,44 @@
+//списки имён, описаний и комментариев
+const NAMES = [
+  'Алеша Попович',
+  'Виктор Самойлов',
+  'Артур Пирожков',
+  'Иван Обрамов',
+  'Иван Охлабыстин',
+  'Клава Кока',
+  'Игорь Крутой',
+  'Иннокентий Кашинский',
+  'Лолита Милявская',
+];
+
+const DESCRIPTIONS_PHOTO = [
+  'Унылая беспечность',
+  'Красота и восторг',
+  'Любо дорого смотреть',
+  'Дело в ньюансах',
+  'Нанавижу этот стиль',
+  'Делоть было нечего, дело было вечером',
+  'Ужасный восторг',
+  'Пятое чудо света',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const MAX_PHOTOS = 25;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 30;
+const MAX_AVATAR_NUMBER = 6;
+
+//Подготовка массива для хранения фотографий
+const photos = [];
+
+export {NAMES,DESCRIPTIONS_PHOTO, COMMENTS, MAX_PHOTOS,MIN_LIKES,MAX_LIKES,MIN_COMMENTS,MAX_COMMENTS,MAX_AVATAR_NUMBER,photos};
